s/a: Extract audit client from manager
