+++ src/v/security/audit/BUILD
@@ -17,6 +17,8 @@ redpanda_cc_library(
     name = "audit",
     srcs = [
         "audit_log_manager.cc",
+        "client.cc",
+        "client.h",  # Maybe split client into its own target
         "logger.cc",
         "probes.cc",
         "schemas/hashing_utils.cc",
@@ -45,6 +47,7 @@ redpanda_cc_library(
         "//src/v/base",
         "//src/v/cluster",
         "//src/v/config",
+        "//src/v/container:chunked_vector",
         "//src/v/hashing:combine",
         "//src/v/json",
         "//src/v/kafka/client",
@@ -104,6 +107,7 @@ redpanda_cc_library(
         "//src/v/ssx:semaphore",
         "//src/v/strings:string_switch",
         "//src/v/utils:functional",
+        "//src/v/utils:mutex",
         "//src/v/utils:retry",
         "//src/v/utils:unresolved_address",
         "//src/v/version",
